backend/pyproject.toml
[tool.poetry]
name = "videoagent"
version = "0.1.0"
description = ""
authors = [
    "Ricky Feng <you@example.com>"
]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
openai = ">=1.76.2,<2.0.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
python-dotenv = "^1.1.0"
google-genai = "^1.13.0"
python-multipart = "^0.0.20"
beautifulsoup4 = "^4.13.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"